requirements.txt
openai>=1.0.0
trimesh
Pillow
numpy
